ensure post publish tasks are run when direct publishing, pull specific functions out of helpers
