+++ api/src/components/affiliations/controller.ts
@@ -5,7 +5,6 @@ import * as response from 'lib/response';
 import * as publicationVersionService from 'publicationVersion/service';
 import * as userService from 'user/service';
 import * as coAuthorService from 'coauthor/service';
-import * as Helpers from 'lib/helpers';
 import { Prisma } from '@prisma/client';
 
 export const updateAffiliations = async (
@@ -83,6 +82,38 @@ export const updateAffiliations = async (
     }
 };
 
+const mapOrcidAffiliationSummary = (
+    summary: I.OrcidAffiliationSummary,
+    affiliationType: I.MappedOrcidAffiliation['affiliationType']
+): I.MappedOrcidAffiliation => ({
+    id: summary['put-code'],
+    affiliationType,
+    title: summary['role-title'],
+    departmentName: summary['department-name'],
+    startDate: summary['start-date']
+        ? {
+              year: summary['start-date'].year?.value || null,
+              month: summary['start-date'].month?.value || null,
+              day: summary['start-date'].day?.value || null
+          }
+        : undefined,
+    endDate: summary['end-date']
+        ? {
+              year: summary['end-date'].year?.value || null,
+              month: summary['end-date'].month?.value || null,
+              day: summary['end-date'].day?.value || null
+          }
+        : undefined,
+    organization: summary.organization,
+    createdAt: summary['created-date'].value,
+    updatedAt: summary['last-modified-date'].value,
+    source: {
+        name: summary.source['source-name'].value,
+        orcid: summary.source['source-orcid']?.path || summary.source['source-client-id']?.path || ''
+    },
+    url: summary.url ? summary.url.value : undefined
+});
+
 export const getOrcidAffiliations = async (
     event: I.AuthenticatedAPIRequest<undefined, undefined, undefined>
 ): Promise<I.JSONResponse> => {
@@ -101,34 +132,33 @@ export const getOrcidAffiliations = async (
         const activitiesSummary = orcidRecordResponse.data['activities-summary'];
 
         const memberships: I.MappedOrcidAffiliation[] = activitiesSummary?.memberships?.['affiliation-group'].map(
-            (item) => Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['membership-summary'], 'membership')
+            (item) => mapOrcidAffiliationSummary(item?.summaries?.[0]?.['membership-summary'], 'membership')
         );
 
         const services: I.MappedOrcidAffiliation[] = activitiesSummary?.services?.['affiliation-group'].map((item) =>
-            Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['service-summary'], 'service')
+            mapOrcidAffiliationSummary(item?.summaries?.[0]?.['service-summary'], 'service')
         );
 
         const invitedPositions: I.MappedOrcidAffiliation[] = activitiesSummary?.['invited-positions']?.[
             'affiliation-group'
         ].map((item) =>
-            Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['invited-position-summary'], 'invited-position')
+            mapOrcidAffiliationSummary(item?.summaries?.[0]?.['invited-position-summary'], 'invited-position')
         );
 
         const distinctions: I.MappedOrcidAffiliation[] = activitiesSummary?.distinctions?.['affiliation-group'].map(
-            (item) => Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['distinction-summary'], 'distinction')
+            (item) => mapOrcidAffiliationSummary(item?.summaries?.[0]?.['distinction-summary'], 'distinction')
         );
 
         const employments: I.MappedOrcidAffiliation[] = activitiesSummary?.employments?.['affiliation-group'].map(
-            (item) => Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['employment-summary'], 'employment')
+            (item) => mapOrcidAffiliationSummary(item?.summaries?.[0]?.['employment-summary'], 'employment')
         );
 
         const educations: I.MappedOrcidAffiliation[] = activitiesSummary?.educations?.['affiliation-group'].map(
-            (item) => Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['education-summary'], 'education')
+            (item) => mapOrcidAffiliationSummary(item?.summaries?.[0]?.['education-summary'], 'education')
         );
 
         const qualifications: I.MappedOrcidAffiliation[] = activitiesSummary?.qualifications?.['affiliation-group'].map(
-            (item) =>
-                Helpers.mapOrcidAffiliationSummary(item?.summaries?.[0]?.['qualification-summary'], 'qualification')
+            (item) => mapOrcidAffiliationSummary(item?.summaries?.[0]?.['qualification-summary'], 'qualification')
         );
 
         return response.json(200, [
